Add modded logs to Reap config (via tag).
This may skip certain naturally generated wood for mods that use the wood variant rather than log variant of a block, so it will need to be adjusted over time to ensure that nothing is missed.

By using a tag, the config file won't need to be updated very often, avoiding a huge serverconfigupdater config in the future (need to investigate if the axes list can use tags too).